Standard implementation finished yipee

+++ src/Frog.cpp
@@ -8,7 +8,7 @@
 #include "FileFormatError.h"
 
 Frog::Frog(PlayState* state, Texture* texture, Point2D<float> position, int lives) :
-	SceneObject(state, texture, position, texture->getFrameWidth(), texture->getFrameHeight()),
+	SceneObject(state, texture, position, static_cast<float>(texture->getFrameWidth()), static_cast<float>(texture->getFrameHeight())),
 	spawnPosition(position),
 	direction(0, 0),
 	lives(lives),
@@ -38,8 +38,8 @@ Frog::Frog(PlayState* state, std::fstream& file, int lineNumber)
 	spawnPosition = position;
 	lives = FrogLives;
     texture = state->getGame()->getTexture(Game::FROG);
-    width = texture->getFrameWidth();
-    height = texture->getFrameHeight();
+    width = static_cast<float>(texture->getFrameWidth());
+    height = static_cast<float>(texture->getFrameHeight());
 }
 
 void Frog::update()
@@ -83,26 +83,26 @@ void Frog::handleEvent(const SDL_Event& event) //direciones de movimiento de la 
 	case SDLK_UP:
 		if (position.getY() > 0)
 		{
-			direction = Vector2D<float>(0, -Game::CellSize);
+			direction = Vector2D<float>(0, -static_cast<float>(Game::CellSize));
 			orientation = 1;
 		}
 		break;
 	case SDLK_DOWN:
 		if (position.getY() + height < Game::WINDOW_HEIGHT)
 		{
-			direction = Vector2D<float>(0, Game::CellSize);
+			direction = Vector2D<float>(0, static_cast<float>(Game::CellSize));
 			orientation = 3;
 		}
 		break;
 	case SDLK_LEFT:
 		if (position.getX() > 0) {
-			direction = Vector2D<float>(-Game::CellSize, 0);
+			direction = Vector2D<float>(-static_cast<float>(Game::CellSize), 0);
 			orientation = 2;
 		}
 		break;
 	case SDLK_RIGHT:
 		if (position.getX() + width < Game::WINDOW_WIDTH) {
-			direction = Vector2D<float>(Game::CellSize, 0);
+			direction = Vector2D<float>(static_cast<float>(Game::CellSize), 0);
 			orientation = 4;
 		}
 		break;
