+++ src/HomedFrog.cpp
@@ -3,7 +3,7 @@
 #include "PlayState.h" // Include PlayState.h
 
 HomedFrog::HomedFrog(PlayState* state, Texture* texture, Point2D<float> pos)
-    : SceneObject(state, texture, pos, texture->getFrameWidth(), texture->getFrameHeight())
+    : SceneObject(state, texture, pos, static_cast<float>(texture->getFrameWidth()), static_cast<float>(texture->getFrameHeight()))
 {
 }
