+++ src/Vehicle.cpp
@@ -8,7 +8,7 @@
 #include "FileFormatError.h"
 
 Vehicle::Vehicle(PlayState* state, Texture* texture, Point2D<float> pos, Vector2D<float> vel)
-    : Crosser(state, texture, pos, texture->getFrameWidth(), texture->getFrameHeight(), vel)
+    : Crosser(state, texture, pos, static_cast<float>(texture->getFrameWidth()), static_cast<float>(texture->getFrameHeight()), vel)
 {
 }
 
@@ -32,8 +32,8 @@ Vehicle::Vehicle(PlayState* state, std::fstream& file, int lineNumber)
     Texture* tex = state->getGame()->getTexture(static_cast<Game::TextureName>(textureIndex));
 
     texture = tex;
-    width = tex->getFrameWidth();
-    height = tex->getFrameHeight();
+    width = static_cast<float>(tex->getFrameWidth());
+    height = static_cast<float>(tex->getFrameHeight());
     position = Point2D<float>(Xpos, Ypos);
     velocity = Vector2D<float>(Xvel / Game::FRAME_RATE, 0.0f);
 }
